third_party/WebKit/LayoutTests/inspector/coverage/gutter-js.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="coverage-test.js"></script>
<script src="resources/coverage.js"></script>

<script>

async function test()
{
    InspectorTest.startCoverage();
    await InspectorTest.evaluateInPagePromise("performActions()");
    await InspectorTest.stopCoverage();
    await InspectorTest.dumpDecorations("coverage.js");
    InspectorTest.completeTest();
}
</script>
</head>
<p id="id">PASS</p>
<body onload="runTest()">
<p>Tests the coverage highlight in sources after the recording finishes.</p>
</body>
</html>
